+++ scopes/component/ui/component-artifacts/artifacts-tree/artifacts-tree.tsx
@@ -14,6 +14,7 @@ import { ComponentID } from '@teambit/component-id';
 import { FolderTreeNode } from '@teambit/ui-foundation.ui.tree.folder-tree-node';
 
 import styles from './artifacts-tree.module.scss';
+import { fileNodeClicked } from './artifact-file-node-clicked';
 
 export type ArtifactsTreeProps = {
   getIcon?: (node: TreeNode) => string | undefined;
@@ -23,31 +24,6 @@ export type ArtifactsTreeProps = {
   componentId: ComponentID;
   drawerName: string;
 } & HTMLAttributes<HTMLDivElement>;
-
-const fileNodeClicked = (files: (ArtifactFile & { id: string })[], opts: 'download' | 'new tab') => (e, node) => {
-  const { id } = node;
-  const artifactFile = files.find((file) => file.id === id);
-
-  if (artifactFile?.downloadUrl) {
-    fetch(artifactFile.downloadUrl, { method: 'GET' })
-      .then((res) => res.blob())
-      .then((blob) => {
-        // create blob link to download
-        const url = window.URL.createObjectURL(new Blob([blob]));
-        const link = document.createElement('a');
-        link.href = url;
-        if (opts === 'download') link.setAttribute('download', artifactFile.path);
-        if (opts === 'new tab') link.setAttribute('target', '_blank');
-        // append to html page
-        document.body.appendChild(link);
-        // force download
-        link.click();
-        // clean up and remove the link
-        link.parentNode?.removeChild(link);
-      })
-      .catch(() => {});
-  }
-};
 
 export function ArtifactsTree({
   getIcon,
@@ -139,7 +115,7 @@ function generateWidget(files: (ArtifactFile & { id: string })[]) {
   };
 }
 
-export function FileTreeNode(props: TreeNodeProps<any>) {
+function FileTreeNode(props: TreeNodeProps<any>) {
   const { node } = props;
   const fileTreeContext = useFileTreeContext();
   const { selected, onSelect } = useContext(TreeContext);
